applying code clean in some files
